feat: added weblayout
